# backend-documentation/openapi.yaml
openapi: 3.0.3
info:
  title: The Game Pensieve API
  description: |
    The Game Pensieve is a comprehensive collection management system for gaming items including toys, video games, board games, and gaming systems. 
    
    This API provides full CRUD operations for managing your gaming collection with support for custom fields, advanced filtering, and data backup/import functionality.

    ## Response Format
    Every **successful** response is wrapped in a consistent envelope:
    ```json
    {
      "data": <response_data>,
      "errors": null,
      "roundTripMs": 12
    }
    ```
    - `data` — the payload (an object, an array, or an empty string `""` for deletes). `null` on error.
    - `errors` — always `null` on success.
    - `roundTripMs` — server-side processing time in milliseconds, measured from request receipt to response. Present only on successful responses produced by the controllers; it is **absent** from error responses.

    ## Error Format
    Errors do **not** include `roundTripMs`. `data` is `null` and `errors` is always an **array of strings** (one entry per problem found — a single request can report several validation errors at once):
    ```json
    {
      "data": null,
      "errors": ["title cannot be blank", "systemId must be a positive integer"]
    }
    ```
    Status codes used:
...
    - **404 Not Found** — the requested resource (or an unknown route) does not exist.
...

    ## Data Ownership & Multi-Tenancy
...

...

...

...

...

...

    ## Timestamps
...

...

...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Log in and obtain tokens
      description: |
        Exchange credentials for a short-lived access token and a longer-lived, revocable refresh token.
        Send the access token as `Authorization: Bearer <accessToken>` on protected routes. Public endpoint:
        no token required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "collector@example.com"
              password: "Sup3rSecret!"
      responses:
        '200':
          description: Authentication succeeded; tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthDataResponse'
              example:
                data:
                  accessToken: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJjb2xsZWN0b3JAZXhhbXBsZS5jb20ifQ.signature"
                  refreshToken: "PHCQ8m1k0t9rJv3w7s2A6bQx4nL5dZ0uHcRfTgYkImE"
                  tokenType: "Bearer"
                  expiresInMs: 900000
                errors: null
                roundTripMs: 18
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                data: null
                errors:
                  - "Invalid email or password."
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
...
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required:
        - email
        - password
...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: "Short-lived JWT to send as `Authorization: Bearer <accessToken>`"
        refreshToken:
          type: string
          description: Longer-lived, single-use, revocable token used to obtain a new access token
        tokenType:
          type: string
          example: "Bearer"
        expiresInMs:
          type: integer
          format: int64
          description: Lifetime of the access token in milliseconds
    AuthDataResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AuthTokens'
        errors:
          type: 'null'
        roundTripMs:
          type: integer
          format: int64
...
    ErrorResponse:
      type: object
      description: |
        Returned for any failed request. `data` is null and `errors` is an array of human-readable
        message strings — one entry per problem detected. Note there is no `roundTripMs` field on errors.
      properties:
        data:
          type: 'null'
        errors:
          type: array
          items:
            type: string
          description: One or more error message strings
      example:
        data: null
        errors:
          - "title cannot be blank"
          - "systemId must be a positive integer"
...